tests/pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["likeinlife <likeinlife@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.3"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
httpx = "^0.25.2"
aenum = "^3.1.15"
PyJWT = {extras = ["crypto"], version = "^2.8.0"}

[tool.poetry.dev-dependencies]
ruff = "^0.1.6"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]
ignore_missing_imports = true

[tool.ruff]
line-length = 120
select = ["D", "B"]
ignore = ["D100", "D107", "D101", "D106", "D103", "D104", "D102"]

